third_party/WebKit/LayoutTests/editing/execCommand/check_justify_center_performance.html

<!doctype html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../assert_selection.js"></script>
<script src=../data/justifycenter.js></script>
<script>
test(() => assert_selection(
  '<div contenteditable><hr><pre>|<hr></pre></div>',
  'JustifyCenter',
  [
  '<div contenteditable>',
    '<hr><pre>|<hr style="text-align: center;"></pre>',
  '</div>'
  ]),'This test is for a crash when performing JustifyCenter.');
</script>
